Add Authorization to the progect

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interface;
 using Restaurants.Domain.Repositories;
 using System;
 using System.Collections.Generic;
@@ -13,7 +15,8 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDishes
 {
 	public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> logger,
 	IRestaurantsRepository restaurantsRepository,
-	IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesForRestaurantCommand>
+	IDishesRepository dishesRepository,
+	IRestaurantAuthorizationServices restaurantAuthorizationServices) : IRequestHandler<DeleteDishesForRestaurantCommand>
 	{
 		public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
 		{
@@ -22,6 +25,9 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDishes
 
 			var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
 			if (restaurant == null) throw new NotfoundException(nameof(Restaurant), request.RestaurantId.ToString());
+
+			if (!restaurantAuthorizationServices.Authorize(restaurant, ResourceOperation.Delete))
+				throw new ForbidException();
 
 			await dishesRepository.Delete(restaurant.Dishes);
